Split AutoConfig into 2 files.
