+++ src/autowiring/AutoConfigBase.cpp
@@ -1,6 +1,6 @@
 // Copyright (C) 2012-2015 Leap Motion, Inc. All rights reserved.
 #include "stdafx.h"
-#include "AutoConfig.h"
+#include "AutoConfigBase.h"
 #include "AutoConfigParser.hpp"
 
 AutoConfigVarBase::AutoConfigVarBase(const std::type_info& ti, bool configured) :
